+++ libnest2d/geometry_traits.hpp
@@ -348,7 +348,7 @@ struct ShapeLike {
 
     // Optional, does nothing by default
     template<class RawShape>
-    static void reserve(RawShape& /*sh*/,  unsigned long /*vertex_capacity*/) {}
+    static void reserve(RawShape& /*sh*/,  size_t /*vertex_capacity*/) {}
 
     template<class RawShape, class...Args>
     static void addVertex(RawShape& sh, Args...args)
